Machine Learning/requirements.txt
Flask>=2.0
flask-cors
requests
beautifulsoup4
python-whois
dnspython
python-dateutil
gunicorn
tldextract
scikit-learn==1.7.2
joblib==1.4.2
pandas==2.2.3
